Divide Dashboard related components
